minor improvements on Footer4
rename footer lucas to Footer4

removing unused dependencies from Footer4
