+++ src/pages/index.tsx
@@ -8,10 +8,8 @@ import Services1 from "@/components/Services/Services1"
 import StatisticsSection from "@/components/Stats/Stats1"
 import Member1 from "@/components/Members/Members1"
 import Clients1 from "@/components/Clients/Clients1"
-import FooterVic from "@/components/Footers/Footer3"
 import Contact1 from "@/components/Contact/Contact1"
-import FooterLucas from "@/components/Footers/footerlucas"
-import Footer1 from "@/components/Footers/Footer1"
+import Footer4 from "@/components/Footers/Footer4"
 
 const inter = Inter({
   subsets: ["latin"],
@@ -35,7 +33,7 @@ const Home: NextPage = () => {
         <Clients1 />
         <Member1 />
         <Contact1 />
-        <FooterLucas/>
+        <Footer4 />
       </main>
     </div>
   )
